converting the project to TypeScript
