+++ .gitignore
@@ -6,3 +6,5 @@ yarn.lock
 .vscode
 .idea
 *.iml
+/lib
+/docs
