+++ .npmignore
@@ -5,3 +5,4 @@ CONTRIBUTING.md
 History.md
 PUBLISHING.md
 .*
+/src
