+++ resources/templates/wsse-security-header.ejs
