+++ resources/templates/wsse-security-token.ejs
