+++ src/utils.ts
@@ -1,0 +1,50 @@
+
+import * as crypto from 'crypto';
+
+export function passwordDigest(nonce: string, created: string, password: string): string {
+  // digest = base64 ( sha1 ( nonce + created + password ) )
+  var pwHash = crypto.createHash('sha1');
+
+  var NonceBytes = Buffer.from(nonce || '', 'base64');
+  var CreatedBytes = Buffer.from(created || '', 'utf8');
+  var PasswordBytes = Buffer.from(password || '', 'utf8');
+  var FullBytes = Buffer.concat([NonceBytes, CreatedBytes, PasswordBytes ]);
+
+  pwHash.update(FullBytes);
+  return pwHash.digest('base64');
+}
+
+export const TNS_PREFIX = '__tns__'; // Prefix for targetNamespace
+
+/**
+ * Find a key from an object based on the value
+ * @param {Object} Namespace prefix/uri mapping
+ * @param {*} nsURI value
+ * @returns {String} The matching key
+ */
+export function findPrefix(xmlnsMapping, nsURI) {
+  for (var n in xmlnsMapping) {
+    if (n === TNS_PREFIX) continue;
+    if (xmlnsMapping[n] === nsURI) {
+      return n;
+    }
+  }
+}
+
+export function splitQName<T>(nsName: T) {
+  if (typeof nsName !== 'string') {
+    return {
+      prefix: TNS_PREFIX,
+      name: nsName,
+    };
+  }
+
+  const [topLevelName] = nsName.split('|');
+
+  const prefixOffset = topLevelName.indexOf(':');
+
+  return {
+    prefix: topLevelName.substring(0, prefixOffset) || TNS_PREFIX,
+    name: topLevelName.substring(prefixOffset + 1),
+  };
+}
