+++ test/client-customHttp-test.js
@@ -6,7 +6,7 @@ var fs = require('fs'),
     assert = require('assert'),
   duplexer = require('duplexer'),
   req = require('request'),
-  httpClient = require('../lib/http.js'),
+  httpClient = require('../lib/http.js').HttpClient,
  // stream = require('stream'),
   stream = require('readable-stream'),
   util = require('util'),
