+++ test/client-customHttp-xsdinclude-test.js
@@ -4,7 +4,7 @@ var soap = require('..'),
     http = require('http'),
     assert = require('assert'),
   req = require('request'),
-  httpClient = require('../lib/http.js'),
+  httpClient = require('../lib/http.js').HttpClient,
   util = require('util'),
   events = require('events'),
   createSocketStream = require('./_socketStream');
